References

Main Structures

PotreeConverter.PotreeWriter — Type

PotreeWriter

Metastructure of potree.

PotreeConverter.PWNode — Type

PWNode

SIngle node of potree.

Main references

PotreeConverter.main — Function

main(sources,outdir,pageName)

Interface.

Input:

sources: list of point clouds (file .las) to convert

outdir: output folder

pageName: project name

Output:

Potree structure saved in outdir: a single file .las for each node.

PotreeConverter.potreeconvert — Function

potreeconvert(args::PotreeArguments)

Generates an octree LOD structure:

Read all points in sources.

Add each point to a one node of octree.

Write all info and structures to the disk.

PotreeConverter.add — Function

add(potreeWriter::PotreeWriter,p::Point)

Add point in PotreeWriter store and process store.

add(sparseGrid::SparseGrid,p::Vector{Float64})::Bool

Add point p in sparse grid. Properties:

compute the index of the cell where p falls

check the distance between p and other points in same cell and in nearest cells

add(node::PWNode, point::Point, potreeWriter::PotreeWriter)::Union{Nothing,PWNode}

Associates the point to one node of the subtree that has node as root. Properties:

the leaf node expands iff the size of store is greater than a chosen threshold.

the internal node accepts points that satisfy requirement.

the point not accepted by this node passes to one of its children

PotreeConverter.processStore — Function

processStore(potreeWriter::PotreeWriter)

Move points from PotreeWriter store in a specific node of octree.