References

PotreeConverter

Main Structures

PotreeConverter.Point — Type

Point

Point information.

Constructors

Point(lasPoint::FileManager.LasPoint, header::FileManager.LasHeader)::Point

Fields

position::Vector{Float64}
color::Vector{UInt16}
normal::Vector{Float64}
intensity::UInt16
classification::Char
returnNumber::Char
numberOfReturns::Char
pointSourceID::UInt16
gpsTime::Float64

PotreeConverter.pAABB — Type

pAABB

Axis Aligned Bounding Box

Constructors

pAABB()
pAABB(min::Vector{Float64},max::Vector{Float64})

Fields

min::Vector{Float64}
max::Vector{Float64}
size::Vector{Float64}

PotreeConverter.PotreeWriter — Type

PotreeWriter

Metastructure of octree.

Constructors

PotreeWriter(workDir::String, aabb::pAABB, root::Union{Nothing,PWNode}, spacing::Float64, maxDepth::Int64, scale::Float64, quality::ConversionQuality )
PotreeWriter(workDir::String, quality::ConversionQuality )

Fields

aabb::pAABB
tightAABB::pAABB
workDir::String
spacing::Float64
scale::Float64
maxDepth::Int64
root::Union{Nothing,PWNode}
numAdded::Int64
numAccepted::Int64
outputFormat::String
pointAttributes::String
hierarchyStepSize::Int
store::Vector{Point}
pointsInMemory::Int
quality::ConversionQuality
storeSize::Int64
storeThread::Union{Nothing,Task}

PotreeConverter.PWNode — Type

PWNode

Single node of octree.

Constructors

PWNode()
PWNode(potreeWriter::PotreeWriter,aabb::pAABB)
PWNode(potreeWriter::PotreeWriter, index::Int, aabb::pAABB, level::Int)

Fields

index::Int
aabb::pAABB
acceptedAABB::pAABB
level::Int
spacing::Float64
grid::SparseGrid
numAccepted::UInt
parent::Union{Nothing,PWNode}
children::Vector{Union{Nothing,PWNode}}
addedSinceLastFlush::Bool
addCalledSinceLastFlush::Bool
cache::Vector{Point}
store::Vector{Point}
isInMemory::Bool

PotreeConverter.SparseGrid — Type

SparseGrid

Sparse grid in a node.

Constructors

SparseGrid(aabb::pAABB, spacing::Float64)

Fields

map::Dict{Int64,GridCell}
width::Int
height::Int
depth::Int
aabb::pAABB
squaredSpacing::Float64
numAccepted::UInt64

Main references

PotreeConverter.main — Function

main(sources,outdir,pageName)

Interface.

Input:

sources: list of point clouds (file .las) to convert

outdir: output folder

pageName: project name

Output:

Potree structure saved in outdir: a single file .las for each node.

PotreeConverter.potreeconvert — Function

potreeconvert(args::PotreeArguments)

Generates an octree LOD structure:

Read all points in sources.

Add each point to a one node of octree.

Write all info and structures to the disk.

PotreeConverter.add — Function

add(potreeWriter::PotreeWriter,p::Point)

Add point in PotreeWriter store and process store.

add(sparseGrid::SparseGrid,p::Vector{Float64})::Bool

Add point p in sparse grid.

Properties:

compute the index of the cell where p falls

check the distance between p and other points in same cell and in nearest cells

add(node::PWNode, point::Point, potreeWriter::PotreeWriter)::Union{Nothing,PWNode}

Associates the point to one node of the subtree that has node as root.

Properties:

the leaf node expands iff the size of store is greater than a chosen threshold.

the internal node accepts points that satisfy requirement.

the point not accepted by this node passes to one of its children

PotreeConverter.processStore — Function

processStore(potreeWriter::PotreeWriter)

Move points from PotreeWriter store in a specific node of octree.

PotreeConverter.flush — Function

flush(potreeWriter::PotreeWriter, cloudjs::CloudJS)

Write to disk all nodes and entire hierarchy.

Properties:

flush all nodes of octree

save the metadata of the octree in cloud.js

write the hierarchy in a file .hrc

flush(node::PWNode, potreeWriter::PotreeWriter)

Write to disk the subtree that has node as root.

Properties:

create a file .las for each node.

Comaptree

Main Structures

PotreeConverter.ComaptreeWriter — Type

ComaptreeWriter

Metastructure of isomorphic tree of Potree, compatree.

Constructors

ComaptreeWriter(root::CWNode)

Fields

root::CWNode

PotreeConverter.CWNode — Type

CWNode

Single node of Comaptree.

Constructors

CWNode(index::Int, level::Int)::CWNode
CWNode(index::Int, level::Int)::CWNode

Fields

level::Int
index::Int
dict::Dict{Vector{Float64},Vector{Vector{Float64}}}
parent::Union{Nothing,CWNode}
children::Vector{Union{Nothing,CWNode}}

Main references

PotreeConverter.potree2comaptree — Method

potree2comaptree(potreeDir::String)

Given a potree, construct an empty isomorphic tree.

PotreeConverter.postorder — Function

postorder(root::PotreeConverter.CWNode, callback::Function)

If the node is not a leaf, then postorder traverse all children of node, else elaborate the node dataset.